+++ src/callbackBuilder.js
@@ -1,12 +1,21 @@
+'use strict';
+
+const toArray = require('./utils').toArray;
+
 /**
  * Build a callback for the given promise resolve/reject functions.
  *
  * @param {Boolean|String[]} options.variadic See the documentation for promisify.
  */
-function callbackBuilder(resolve, reject, {variadic}) {
+function callbackBuilder(resolve, reject, options) {
+  let variadic;
+  if (options) {
+    variadic = options.variadic;
+  }
+
   let called = false;
 
-  return function callback(err, ...values) {
+  return function callback(err, value) {
     if (called) {
       throw new Error('the deferred callback has already been called');
     }
@@ -18,11 +27,11 @@ function callbackBuilder(resolve, reject, {variadic}) {
     } else if (Array.isArray(variadic)) {
       const obj = {};
       for (let i = 0; i < variadic.length; i++) {
-        obj[variadic[i]] = values[i];
+        obj[variadic[i]] = arguments[i + 1];
       }
       resolve(obj);
     } else {
-      resolve(variadic ? values : values[0]);
+      resolve(variadic ? toArray(arguments, 1) : value);
     }
   };
 }