+++ src/statics/delay.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function delay(time, value) {
+  return new Promise((resolve) => setTimeout(resolve, time, value));
+}
+
+module.exports = delay;