+++ src/statics/immediate.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function immediate(value) {
+  return new Promise((resolve) => setImmediate(resolve, value));
+}
+
+module.exports = immediate;