+++ src/statics/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+module.exports = [
+  require('./asCallback'),
+  require('./delay'),
+  require('./immediate'),
+  require('./nextTick'),
+  require('./waitOn'),
+  require('./withTimeout'),
+  require('./wrapAsync'),
+];