+++ src/statics/nextTick.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function nextTick(value) {
+  return new Promise((resolve) => process.nextTick(resolve, value));
+}
+
+module.exports = nextTick;