+++ src/utils.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const arraySlice = Array.prototype.slice;
+
+function toArray(arrayLike, offset) {
+  return arraySlice.call(arrayLike, offset);
+}
+
+module.exports = {
+  toArray,
+};