+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Release History
 
+* 3.0.0 Add waitOn, wrapAsync
+        Remove copyAll option from promisifyAll
+        Fix support for Node 4
 * 2.1.3 Fix delay, static exports, accept resolve value for async primitives
 * 2.1.2 Fix for Node < 7
 * 2.0.0 Replace sync with deferred, add promisify, fix asCallback